requirements.txt
# sudo apt-get install graphviz libgraphviz-dev
# sudo apt-get install python3-tk
rcviz
pillow
matplotlib